amd64: shrink operand storage metadata
Store validated operand indexes and bounded GC payloads in 32 bits, and pack GC/EH root state with value facts behind preserving accessors. This reduces storage from 32 to 24 bytes and elem from 72 to 64 bytes without changing generated code.\n\nEmulated Linux/AMD64 json-as compilation drops 33,632 B/op (12.2%) and many_funcs drops 640 B/op with allocation counts unchanged. The complete Linux/AMD64 suite passes and representative code hashes are identical.

+++ src/core/compiler/backend/railshot/amd64/globals.go
@@ -74,7 +74,7 @@ func (f *fn) globalGet(r *wasm.Reader) error {
 		if wasm.EqualValType(gtv, wasm.I64) {
 			typ = mtI64
 		}
-		f.pushValue(storage{kind: stGlobReg, typ: typ, reg: reg, idx: int(x)})
+		f.pushValue(storage{kind: stGlobReg, typ: typ, reg: reg, idx: x})
 		return nil
 	}
 	cell := f.globalCellPtr(x) // cached, pinned — read the value into a separate reg
@@ -116,7 +116,7 @@ func (f *fn) realizeGlobalRefs(x uint32, skipFrom *elem) {
 		}
 		next := e.next
 		switch {
-		case e.kind == ekValue && e.st.kind == stGlobReg && uint32(e.st.idx) == x:
+		case e.kind == ekValue && e.st.kind == stGlobReg && e.st.idx == x:
 			f.materialize(e)
 		case e.kind == ekDeferred && subtreeRefsGlobal(e, x):
 			f.condense(e, regNone)
@@ -132,7 +132,7 @@ func subtreeRefsGlobal(e *elem, x uint32) bool {
 		return false
 	}
 	if e.kind == ekValue {
-		return e.st.kind == stGlobReg && uint32(e.st.idx) == x
+		return e.st.kind == stGlobReg && e.st.idx == x
 	}
 	if e.kind == ekDeferred {
 		return subtreeRefsGlobal(e.arg0, x) || subtreeRefsGlobal(e.arg1, x)
